Introduced gateway API made more sense from config and listener point of view.
